De-bloat MatchManager, add match summary
